FineLIP-main/FineLIP-main/requirements.txt
absl-py==2.1.0
accelerate==0.29.3
aiobotocore==2.12.3
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==0.7.16
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
botocore==1.34.69
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.0
dill==0.3.8
diffusers==0.30.2
docker-pycreds==0.4.0
docstring_parser==0.16
docutils==0.16
eval_type_backport==0.2.0
filelock==3.14.0
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
futures==3.0.5
gast==0.5.4
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.5
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
huggingface-hub==0.24.7
idna==3.7
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
invisible_watermark==0.2.0
jinja2==3.1.3
jmespath==1.0.1
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
namex==0.0.8
numpy==1.26.4
open_clip_torch==2.26.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycocoevalcap
Pygments==2.17.2
pyparsing==3.1.2
pysocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
pyyaml==6.0.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
s3==3.0.0
s3fs==2024.3.1
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
sentry-sdk==2.0.1
setproctitle==1.3.3
setuptools==69.5.1
shtab==1.7.1
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==5.3.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.1
triton==2.2.0
trl==0.8.6
typing_extensions==4.11.0
tyro==0.8.3
tzdata==2024.1
urllib3==1.26.18
wandb==0.16.6
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.16.0
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
ffmpeg==1.4